// frontend/src/app/kupovina/putno-osiguranje/types.ts
export interface FormData {
  whoTravels: string;
  purpose: string;
  selectedCountry: string;
  startDate: string;
  endDate: string;
  birthDay: string;
  birthMonth: string;
  birthYear: string;
  adults: number;
  children: number;
  seniors: number;
  coverage: string;
  termsAccepted: boolean;
}